Module reorganization and doc revisions
